backend/requirements.txt
# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.30.6

# 数据库（MySQL 驱动）
sqlalchemy==2.0.35
PyMySQL==1.1.1
cryptography==43.0.1

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# AI / LLM 集成
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0

# 工具
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2

# 测试
pytest==8.3.3
hypothesis==6.112.2
pytest-asyncio==0.24.0
